+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/stars/StarTransaction.kt
@@ -1,5 +1,7 @@
 package eu.vendeli.tgbot.types.stars
 
+import eu.vendeli.tgbot.utils.serde.InstantSerializer
+import kotlinx.datetime.Instant
 import kotlinx.serialization.Serializable
 
 /**
@@ -16,7 +18,8 @@ import kotlinx.serialization.Serializable
 data class StarTransaction(
     val id: String,
     val amount: Int,
-    val date: Int,
+    @Serializable(InstantSerializer::class)
+    val date: Instant,
     val source: TransactionPartner? = null,
     val receiver: TransactionPartner? = null,
 )