+++ telegram-bot/src/jsMain/kotlin/eu/vendeli/tgbot/utils/BotUtils.js.kt
@@ -1,19 +1,10 @@
 package eu.vendeli.tgbot.utils
 
 import eu.vendeli.tgbot.TelegramBot
-import eu.vendeli.tgbot.interfaces.InputListener
-import eu.vendeli.tgbot.types.User
-import eu.vendeli.tgbot.types.internal.ChainLink
 import kotlinx.coroutines.Dispatchers
 import kotlin.reflect.KClass
 
 internal actual val PROCESSING_DISPATCHER = Dispatchers.Unconfined
-
-@Suppress("NOTHING_TO_INLINE")
-actual inline fun <T : ChainLink> InputListener.setChain(
-    user: User,
-    firstLink: T,
-) = set(user, firstLink::class.simpleName!!)
 
 fun TelegramBot.defineActivities(input: Map<String, List<Any?>>) {
     activities = input